requirements.txt.txt
Django==5.0.7               
djangorestframework==3.15.1 
djangorestframework-simplejwt==5.3.1  
Pillow==10.3.0               
django-filter==24.2          
psycopg2-binary==2.9.9        
python-decouple==3.8 